support read config file path from env CLIENT_CONFIG_FILE_PATH (#76)
* support read config file path from env CLIENT_CONFIG_FILE_PATH

+++ packages/mystiko-contracts/deploy/common.js
@@ -10,16 +10,32 @@ module.exports = {
   RED: RED,
 
   readJsonFile(fileName) {
-    const data = fs.readFileSync(fileName);
-    return JSON.parse(data);
+    if (!fs.existsSync(fileName)) {
+      console.error(this.RED, fileName, ' not exist');
+      return null;
+    }
+
+    try {
+      const data = fs.readFileSync(fileName);
+      return JSON.parse(data);
+    } catch (err) {
+      console.error(this.RED, err);
+      console.error(this.RED, 'read file error');
+    }
   },
 
   writeJsonFile(fileName, data) {
     let jsonData = JSON.stringify(data, null, 2);
     try {
+      if (!fs.existsSync(fileName)) {
+        console.error(this.RED, fileName, ' not exist');
+        return;
+      }
+
       fs.writeFileSync(fileName, jsonData);
     } catch (err) {
-      console.error(err);
+      console.error(this.RED, err);
+      console.error(this.RED, 'write file error');
     }
   },
 
@@ -47,6 +63,14 @@ module.exports = {
       console.error(this.RED, 'save base address config network not support');
       return;
     }
+  },
+
+  resetDefaultDevelopmentConfig(config) {
+    config.chains[0].hashAddress = '';
+    config.chains[0].verifierAddress = '';
+    config.bridges[0].pairs[0][0].address = '';
+
+    this.saveConfig('development', config);
   },
 
   saveBaseAddressConfig(mystikoNetwork, network, config, hashAddress, verifierAddress) {
@@ -106,7 +130,7 @@ module.exports = {
       }
     }
 
-    return null;
+    return '';
   },
 
   getBridgePairIndexByTokenName(bridge, srcNetwork, dstNetwork, tokenName) {
